back/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.61.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
oauthlib==3.2.2
openai==1.30.1
pillow==10.3.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
